DAS-2348: Move CLI entry into regridding_utility
