+++ harmony_regridding_service/regridding_service.py
@@ -32,57 +32,12 @@ from harmony_regridding_service.exceptions import (
     RegridderException,
     SourceDataError,
 )
-from harmony_regridding_service.regridding_utility import (
-    get_harmony_message_from_params,
-)
 
 logger = getLogger(__name__)
 
 HRS_VARINFO_CONFIG_FILENAME = str(
     Path(Path(__file__).parent, 'config', 'HRS_varinfo_config.json')
 )
-
-
-def regrid_cli_entry(
-    source_filename: str, params: dict, source: dict, call_logger: Logger
-):
-    """Call regrid without the adapter class.
-
-    TODO: This is where a library entry will exist if it is ever made.  In the
-    meantime, this is an entrypoint to call the regridder without instantiating
-    an adapter for testing.
-
-    Args:
-    source_filename: [str], a string to an input file to be regridded.
-
-    params: [dict | None], A dictionary with the following keys:
-        crs: [dict | None], Target image's Coordinate Reference System.
-             A dictionary with 'epsg', 'proj4' or 'wkt' key.
-
-        scale_extent: [dict | None], Scale Extents for the image. This dictionary
-            contains "x" and "y" keys each whose value which is a dictionary
-            of "min", "max" values in the same units as the crs.
-            e.g.: { "x": { "min": 0.5, "max": 125 },
-                    "y": { "min": 52, "max": 75.22 } }
-
-        scale_size: [dict | None], Scale sizes for the image.  The dictionary
-            contains "x" and "y" keys with the horizontal and veritcal
-            resolution in the same units as the crs.
-            e.g.: { "x": 10, "y": 10 }
-
-        height: [int | None], height of the output image in gridcells.
-
-        width: [int | none], width of the output image in gridcells.
-
-    source: [dict | None], a Dictionary suitable for initializing a Harmony
-            Source JsonObject.
-
-    call_logger: [Logger], a configured logging object.
-
-    """
-    harmony_message = get_harmony_message_from_params(params)
-    source = HarmonySource(source)
-    return regrid(harmony_message, source_filename, source, call_logger)
 
 
 def regrid(
@@ -789,11 +744,30 @@ def _cache_resamplers(
                 dimensions, filepath, var_info, variable_set
             )
             grid_cache[dimensions] = DaskEWAResampler(source_swath, target_area)
-
-    for resampler in grid_cache.values():
-        resampler.precompute(rows_per_scan=0)
+            rows_per_scan = _get_rows_per_scan(source_swath.shape[0])
+            grid_cache[dimensions].precompute(
+                rows_per_scan=rows_per_scan, maximum_weight_mode=True
+            )
 
     return grid_cache
+
+
+def _get_rows_per_scan(total_rows: int) -> int:
+    """Gets optimum value for rows per scan.
+
+    Finds the smallest divisor of the total number of rows. If no divisor is
+    found, return the total number of rows.
+
+    """
+    if total_rows < 2:
+        return 1
+    for row_number in range(2, int(total_rows**0.5) + 1):
+        if total_rows % row_number == 0:
+            logger.info(f'rows_per_scan = {row_number}')
+            return row_number
+
+    logger.info(f'returning all rows for rows_per_scan = {total_rows}')
+    return total_rows
 
 
 def _compute_target_area(message: HarmonyMessage) -> AreaDefinition:
